+++ src/main/java/TaskList.java
@@ -1,32 +1,13 @@
 import java.util.ArrayList;
-import java.io.BufferedReader;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 
 public class TaskList {
 
     private final ArrayList<Task> myList;
 
-    TaskList() {
-        this.myList = new ArrayList<>();
-        try {
-            BufferedReader br = new BufferedReader(new FileReader("duke.txt"));
-            String line = br.readLine();
-            while (line != null) {
-                textToObject(line);
-                line = br.readLine();
-            }
-            br.close();
-        } catch (FileNotFoundException e){
-            System.out.println("Existing text file not found, creating new file! :)");
-        } catch (IOException e) {
-            System.out.println("IO Exception Error.");
-            e.printStackTrace();
-        }
+    TaskList(ArrayList<Task> myList) {        
+        this.myList = myList;
     }
 
     void addTask(Task myTask) {
@@ -37,33 +18,6 @@ public class TaskList {
             System.out.println("Now you have " + myList.size() + " task(s) in the list.");
         } catch (Exception e) {
             System.out.println("☹ OOPS!!! The description of a todo cannot be empty.");
-        }
-    }
-
-    private LocalDate dateFormatter(String myDate) {
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMM d yyyy");
-        LocalDate curDate = LocalDate.parse(myDate, formatter);
-        return curDate;
-    }
-
-    void textToObject(String line) {
-        String taskType = Character.toString(line.charAt(1));
-        switch (taskType){
-            case "T":
-                myList.add(new Todo(line.substring(7)));
-                break;
-            case "D":
-                myList.add(new Deadline(line.substring(7).split(" \\(by")[0], 
-                    this.dateFormatter(line.split("\\(by: ")[1].split("\\)")[0]))
-                );
-                break;
-            case "E":
-                myList.add(new Event(line.substring(7).split(" \\(at")[0], 
-                    this.dateFormatter(line.split("\\(at: ")[1].split("\\)")[0]))
-                );
-                break;
-            default:
-                break;
         }
     }
